+++ src/main/java/com/debug/api/service/user/UserRefreshTokenService.java
@@ -1,4 +1,4 @@
-package com.debug.api.service;
+package com.debug.api.service.user;
 
 import com.debug.api.entity.user.User;
 import com.debug.api.entity.user.UserRefreshToken;
