+++ src/test/java/com/debug/docs/auth/AuthControllerTests.java
@@ -7,13 +7,12 @@ import com.debug.api.entity.user.User;
 import com.debug.api.entity.user.UserRefreshToken;
 import com.debug.api.exception.LoginFailedException;
 import com.debug.api.repository.user.UserRepository;
-import com.debug.api.service.UserRefreshTokenService;
+import com.debug.api.service.user.UserRefreshTokenService;
 import com.debug.config.properties.AppProperties;
 import com.debug.oauth.entity.RoleType;
 import com.debug.oauth.entity.UserPrincipal;
 import com.debug.oauth.token.AuthToken;
 import com.debug.oauth.token.AuthTokenProvider;
-import com.debug.util.CookieUtil;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.jsonwebtoken.security.Keys;
 import org.junit.jupiter.api.BeforeEach;
